+++ build.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if [ ! -f .cf-key ]; then
+    echo ".cf-key missing!"
+    exit 1
+fi
+
 rm ./*.zip || echo "No ZIPs to delete"
 rm -rf .target || echo "No .target/ to delete"
 mkdir .target
@@ -14,10 +19,14 @@ python3.9 -m pip install \
     cryptography
 
 cp ./*.py .target/
-cp .env.* .target/
+cp .env.shared .target/
 
-# MAIN_CSS_HASH=($(md5sum assets/main.css))
-# echo "MAIN_CSS_HASH=${MAIN_CSS_HASH}" >> .target/.env.shared
+MAIN_CSS_HASH=($(md5sum assets/main.css))
+MAIN_JS_HASH=($(md5sum assets/main.js))
+echo "MAIN_CSS_HASH=${MAIN_CSS_HASH}" >> .target/.env.shared
+echo "MAIN_JS_HASH=${MAIN_JS_HASH}" >> .target/.env.shared
+
+echo "AWS_CLOUDFRONT_KEY=$(cat .cf-key)" >> .target/.env.shared
 
 cp -r templates/ .target/
 cp -r assets/ .target/
